Changed name from AddNodeFix to BetterNodeCheckFix
